Use a random port for WebAgg tests
This prevents them from conflicting when run in parallel. Even though
there is some attempt to pick alternative ports, that seems to cause a
socket leak somewhere in Tornado, which Pytest converts to an error, so
we have to do it manually.

Also ensure the interactive WebAgg subprocess is actually killed.